Move the chart component
